backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
pymongo==4.6.0
chromadb==0.4.22
llama-index==0.10.68
llama-index-llms-openai==0.1.27
llama-index-embeddings-openai==0.1.11
llama-index-readers-file==0.1.6
llama-index-vector-stores-chroma==0.1.9
pymupdf==1.23.26
dappier==0.3.6
